+++ openghg/retrieve/_access.py
@@ -27,6 +27,7 @@ multDataTypes = Union[
 def _get_generic(
     combine_multiple_inlets: bool = False,
     ambig_check_params: list | None = None,
+    version: str = "latest",
     **kwargs: Any,
 ) -> _BaseData:
     """Perform a search and create a dataclass object with the results if any are found.
@@ -36,6 +37,7 @@ def _get_generic(
         combine_multiple_inlets: if multiple results are found, combine them and elevate inlet
             to a data variable.
         ambig_check_params: Parameters to check and print if result is ambiguous.
+        version: Version of data to retrieve. Default = "latest".
         kwargs: Additional search terms
 
     Returns:
@@ -52,7 +54,7 @@ def _get_generic(
         raise SearchError(err_msg)
 
     # TODO: UPDATE THIS - just use retrieve when retrieve_all is removed.
-    retrieved_data = results.retrieve_all()
+    retrieved_data = results.retrieve_all(version=version)
 
     if retrieved_data is None:
         err_msg = f"Unable to retrieve results for {keyword_string}"
@@ -104,6 +106,7 @@ def get_obs_surface(
     keep_variables: list | None = None,
     target_units: dict | None = None,
     is_dequantified: bool = True,
+    version: str = "latest",
     **kwargs: Any,
 ) -> ObsData | None:
     """This is the equivalent of the get_obs function from the ACRG repository.
@@ -134,6 +137,7 @@ def get_obs_surface(
             "mf_variability": "ppm"
         }
         is_dequantified: To dequantify the dataset after getting assigned with pint units. By default it will dequantify the data upon return. To keep the quantification applied supply `False`.
+        version: Version of data to retrieve. Default = "latest".
         kwargs: Additional search terms
 
     Returns:
@@ -179,6 +183,7 @@ def get_obs_surface(
     retrieved_data = _get_generic(
         combine_multiple_inlets=isinstance(inlet, slice),  # if range passed for inlet, try to combine
         ambig_check_params=["inlet", "network", "instrument"],
+        version=version,
         **surface_keywords,  # type: ignore
     )
 
@@ -271,6 +276,7 @@ def get_obs_column(
     average: str | None = None,
     target_units: dict | None = None,
     is_dequantified: bool = True,
+    version: str = "latest",
     **kwargs: Any,
 ) -> ObsColumnData:
     """Extract available column data from the object store using keywords.
@@ -298,11 +304,12 @@ def get_obs_column(
             "mf_variability": "ppm"
         }
         is_dequantified: To dequantify the dataset after getting assigned with pint units. By default it will dequantify the data upon return. To keep the quantification applied supply `False`.
+        version: Version of data to retrieve. Default = "latest".
         kwargs: Additional search terms
     Returns:
         ObsColumnData: ObsColumnData object
     """
-    column_keywords = dict(
+    obs_data = _get_generic(
         species=species,
         satellite=satellite,
         domain=domain,
@@ -314,14 +321,15 @@ def get_obs_column(
         start_date=start_date,
         end_date=end_date,
         data_type="column",
+        **kwargs,
     )
-    column_keywords.update(kwargs)
+    # column_keywords.update(kwargs)
 
-    obs_data = _get_generic(**column_keywords)
+    # obs_data = _get_generic(**column_keywords)
 
     # check if data set is empty
     if obs_data.data.sizes["time"] == 0:
-        raise SearchError(f"Dataset is empty for obs. with {column_keywords}.")
+        raise SearchError("Dataset is empty for obs. Please check the supplied args.")
 
     if return_mf:
         if max_level > max(obs_data.data.lev.values) + 1:
@@ -422,6 +430,7 @@ def get_flux(
     time_resolution: str | None = None,
     target_units: dict | None = None,
     is_dequantified: bool = True,
+    version: str = "latest",
     **kwargs: Any,
 ) -> FluxData:
     """The flux function reads in all flux files for the domain and species as an xarray Dataset.
@@ -442,6 +451,7 @@ def get_flux(
             "mf_variability": "ppm"
         }
         is_dequantified: To dequantify the dataset after getting assigned with pint units. By default it will dequantify the data upon return. To keep the quantification applied supply `False`.
+        version: Version of data to retrieve. Default = "latest".
         kwargs: Additional search terms
     Returns:
         FluxData: FluxData object
@@ -457,6 +467,7 @@ def get_flux(
         start_date=start_date,
         end_date=end_date,
         data_type="flux",
+        version=version,
         **kwargs,
     )
 
@@ -481,6 +492,7 @@ def get_bc(
     end_date: str | Timestamp | None = None,
     target_units: dict | None = None,
     is_dequantified: bool = True,
+    version: str = "latest",
     **kwargs: Any,
 ) -> BoundaryConditionsData:
     """Get boundary conditions for a given species, domain and bc_input name.
@@ -500,6 +512,7 @@ def get_bc(
             "mf_variability": "ppm"
         }
         is_dequantified: To dequantify the dataset after getting assigned with pint units. By default it will dequantify the data upon return. To keep the quantification applied supply `False`.
+        version: Version of data to retrieve. Default = "latest".
     Returns:
         BoundaryConditionsData: BoundaryConditionsData object
     """
@@ -510,6 +523,7 @@ def get_bc(
         start_date=start_date,
         end_date=end_date,
         data_type="boundary_conditions",
+        version=version,
         **kwargs,
     )
 
@@ -530,6 +544,7 @@ def get_footprint(
     species: str | None = None,
     target_units: dict | None = None,
     is_dequantified: bool = True,
+    version: str = "latest",
     **kwargs: Any,
 ) -> FootprintData:
     """Get footprints from one site.
@@ -559,6 +574,7 @@ def get_footprint(
             "mf_variability": "ppm"
         }
         is_dequantified: To dequantify the dataset after getting assigned with pint units. By default it will dequantify the data upon return. To keep the quantification applied supply `False`.
+        version: Version of data to retrieve. Default = "latest".
         kwargs: Additional search terms
     Returns:
         FootprintData: FootprintData dataclass
@@ -586,6 +602,7 @@ def get_footprint(
         end_date=end_date,
         species=species,
         data_type="footprints",
+        version=version,
         **kwargs,
     )