+++ openghg/transform/flux/_edgar.py
@@ -65,7 +65,6 @@ from openghg.util import (
     timestamp_now,
 )
 
-
 logger = logging.getLogger("openghg.transform.flux")
 logger.setLevel(logging.DEBUG)  # Have to set level for logger as well as handler
 
@@ -73,7 +72,7 @@ logger.setLevel(logging.DEBUG)  # Have to set level for logger as well as handle
 ArrayType = Optional[ndarray | xr.DataArray]
 
 
-_edgar_known_versions = ("v432", "v50", "v60", "v70", "v80")
+_edgar_known_versions = ("v432", "v4.3.2", "v50", "v5.0", "v60", "v6.0", "v70", "v7.0", "v80", "v8.0")
 
 
 # TODO: make this work for...