+++ openghg/types/_enum.py
@@ -72,6 +72,12 @@ class EulerianModelTypes(Enum):
     OPENGHG = "OPENGHG"
 
 
+class MetTypes(Enum):
+    """For standardising meteorological data"""
+
+    ECMWF = "ECMWF"
+
+
 class DataTypes(Enum):
     CRDS = "CRDS"
     GCWERKS = "GCWERKS"