+++ tests/storage/test_store.py
@@ -18,22 +18,6 @@ from openghg.types import DataOverlapError
 from openghg.util._versioning import SimpleVersioning, VersionError
 
 
-# idx1 = pd.date_range("2020-01-01", "2020-01-02", freq="h", inclusive="left")
-# idx2 = idx1[:5].union(idx1[10:])  # missing indices 5, 6, 7, 8, 9
-# idx3 = idx1[:5].union(idx1[5:10] + pd.Timedelta("1m")).union(idx1[10:])  # modified indices 5, 6, 7, 8, 9
-# idx4 = idx1 + pd.Timedelta("4h")  # shift by 4 hours
-# idx5 = pd.date_range("2020-01-02", "2020-01-03", freq="h", inclusive="left")  # index starting after idx1
-
-# ds1 = dummy_dataset(idx1)
-# ds2 = dummy_dataset(idx2)
-# ds3 = dummy_dataset(idx3)
-# ds4 = dummy_dataset(idx4)
-# ds5 = dummy_dataset(idx5)
-
-# # test data for updating
-# twice_ds1 = dummy_dataset(idx1).map(lambda x: 2 * x)
-# twice_ds2 = dummy_dataset(idx2).map(lambda x: 2 * x)  # non-contiguous with idx1
-
 # DATA FIXTURES
 def dummy_dataset(index, data=None) -> xr.Dataset:
     if data is None:
@@ -149,19 +133,24 @@ store_names = [
     "versioned_zarr_memory_store",
     "versioned_zarr_directory_store",
 ]
-store_names_is_versioned = [(name, "versioned" in name) for name in store_names]
 
 
+# ------------------------------
 # TESTS
+# ------------------------------
 
 
 # To use fixtures in parametrize, use the "request" fixture, as detailed here:
 # https://stackoverflow.com/questions/42014484/pytest-using-fixtures-as-arguments-in-parametrize
-@pytest.mark.parametrize("store_name, is_versioned", store_names_is_versioned)
-def test_insert_creates(store_name, is_versioned, request, ds1):
+@pytest.mark.parametrize("store_name", store_names)
+def test_insert_creates(store_name, request, ds1):
+    """Test that a store registers as False until data is added.
+
+    After data is added, we check that we can retrieve the data added.
+    """
     store = request.getfixturevalue(store_name)
 
-    if is_versioned:
+    if isinstance(store, VersionedStore):
         store.create_version("v1", checkout=True)
 
     assert not store
@@ -173,11 +162,12 @@ def test_insert_creates(store_name, is_versioned, request, ds1):
     xr.testing.assert_equal(store.get(), ds1)
 
 
-@pytest.mark.parametrize("store_name, is_versioned", store_names_is_versioned)
-def test_clear(store_name, is_versioned, request, ds1):
+@pytest.mark.parametrize("store_name", store_names)
+def test_clear(store_name, request, ds1):
+    """Test clearing data from the store."""
     store = request.getfixturevalue(store_name)
 
-    if is_versioned:
+    if isinstance(store, VersionedStore):
         store.create_version("v1", checkout=True)
 
     assert not store
@@ -192,11 +182,17 @@ def test_clear(store_name, is_versioned, request, ds1):
     xr.testing.assert_identical(store.get(), xr.Dataset())
 
 
-@pytest.mark.parametrize("store_name, is_versioned", store_names_is_versioned)
-def test_insert_twice(store_name, is_versioned, request, ds1, ds5):
+@pytest.mark.parametrize("store_name", store_names)
+def test_insert_twice(store_name, request, ds1, ds5):
+    """Test inserting data twice.
+
+    The datasets `ds1` and `ds5` do not overlap, so after
+    inserting both datasets, the data stored should just be
+    a concatenation of the values of each dataset.
+    """
     store = request.getfixturevalue(store_name)
 
-    if is_versioned:
+    if isinstance(store, VersionedStore):
         store.create_version("v1", checkout=True)
 
     store.insert(ds1)
@@ -207,12 +203,12 @@ def test_insert_twice(store_name, is_versioned, request, ds1, ds5):
     np.testing.assert_equal(store.get().x.values, expected)
 
 
-@pytest.mark.parametrize("store_name, is_versioned", store_names_is_versioned)
-def test_error_on_insert_overlap(store_name, is_versioned, request, ds1):
+@pytest.mark.parametrize("store_name", store_names)
+def test_error_on_insert_overlap(store_name, request, ds1):
     """Test an error is raised on overlap."""
     store = request.getfixturevalue(store_name)
 
-    if is_versioned:
+    if isinstance(store, VersionedStore):
         store.create_version("v1", checkout=True)
 
     store.insert(ds1)
@@ -221,12 +217,12 @@ def test_error_on_insert_overlap(store_name, is_versioned, request, ds1):
         store.insert(ds1)
 
 
-@pytest.mark.parametrize("store_name, is_versioned", store_names_is_versioned)
-def test_insert_ignore_overlap(store_name, is_versioned, request, ds1, ds4):
+@pytest.mark.parametrize("store_name", store_names)
+def test_insert_ignore_overlap(store_name, request, ds1, ds4):
     """Test that insert with `on_overlap = 'ignore'` inserts non-overlaping values."""
     store = request.getfixturevalue(store_name)
 
-    if is_versioned:
+    if isinstance(store, VersionedStore):
         store.create_version("v1", checkout=True)
 
     store.insert(ds1)
@@ -243,11 +239,11 @@ def test_insert_ignore_overlap(store_name, is_versioned, request, ds1, ds4):
     np.testing.assert_equal(store.get().x.values, expected)
 
 
-@pytest.mark.parametrize("store_name, is_versioned", store_names_is_versioned)
-def test_update(store_name, is_versioned, request, ds1, twice_ds1):
+@pytest.mark.parametrize("store_name", store_names)
+def test_update(store_name, request, ds1, twice_ds1):
     store = request.getfixturevalue(store_name)
 
-    if is_versioned:
+    if isinstance(store, VersionedStore):
         store.create_version("v1", checkout=True)
 
     store.insert(ds1)
@@ -260,12 +256,12 @@ def test_update(store_name, is_versioned, request, ds1, twice_ds1):
     xr.testing.assert_equal(store.get(), twice_ds1)
 
 
-@pytest.mark.parametrize("store_name, is_versioned", store_names_is_versioned)
-def test_update_ignore_nonoverlaps(store_name, is_versioned, request, ds1, ds4):
+@pytest.mark.parametrize("store_name", store_names)
+def test_update_ignore_nonoverlaps(store_name, request, ds1, ds4):
     """Test `update` with non-overlaps ignored."""
     store = request.getfixturevalue(store_name)
 
-    if is_versioned:
+    if isinstance(store, VersionedStore):
         store.create_version("v1", checkout=True)
 
     store.insert(ds1)
@@ -279,10 +275,8 @@ def test_update_ignore_nonoverlaps(store_name, is_versioned, request, ds1, ds4):
     np.testing.assert_equal(store.get().x.values, expected)
 
 
-@pytest.mark.parametrize(
-    "store_name, is_versioned, is_zarr", [(name, "versioned" in name, "zarr" in name) for name in store_names]
-)
-def test_non_contiguous_update(store_name, is_versioned, is_zarr, request, ds1, twice_ds2):
+@pytest.mark.parametrize("store_name", store_names)
+def test_non_contiguous_update(store_name, request, ds1, twice_ds2):
     """Test `update` when only some of the values are updated.
 
     This raises an error with Zarr stores because the region to update is non-contiguous.
@@ -290,7 +284,7 @@ def test_non_contiguous_update(store_name, is_versioned, is_zarr, request, ds1, 
     """
     store = request.getfixturevalue(store_name)
 
-    if is_versioned:
+    if isinstance(store, VersionedStore):
         store.create_version("v1", checkout=True)
 
     store.insert(ds1)
@@ -298,23 +292,19 @@ def test_non_contiguous_update(store_name, is_versioned, is_zarr, request, ds1, 
     xr.testing.assert_equal(store.get(), ds1)
 
     # update the values in a non-contiguous region
-    if not is_zarr:
-        store.update(twice_ds2)
+    store.update(twice_ds2)
 
-        expected = np.hstack([twice_ds2.x.values[:5], ds1.x.values[5:10], twice_ds2.x.values[5:]])
+    expected = np.hstack([twice_ds2.x.values[:5], ds1.x.values[5:10], twice_ds2.x.values[5:]])
 
-        np.testing.assert_equal(store.get().x.values, expected)
-    else:
-        with pytest.raises(NotImplementedError):
-            store.update(twice_ds2)
+    np.testing.assert_equal(store.get().x.values, expected)
 
 
-@pytest.mark.parametrize("store_name, is_versioned", store_names_is_versioned)
-def test_contiguous_update(store_name, is_versioned, request, ds1, twice_ds1, ds5):
+@pytest.mark.parametrize("store_name", store_names)
+def test_contiguous_update(store_name, request, ds1, twice_ds1, ds5):
     """Test `update` on contiguous region of times."""
     store = request.getfixturevalue(store_name)
 
-    if is_versioned:
+    if isinstance(store, VersionedStore):
         store.create_version("v1", checkout=True)
 
     store.insert(ds1)